LFU-973-1 mipi_dsi: widen data lane timing fields to u32
Extend data_hs2lp and data_lp2hs in struct mipi_dsi_phy_timing from u16
to u32 to prevent overflow when PHY timing values exceed the 16-bit
range.

+++ include/mipi_dsi.h
@@ -104,8 +104,8 @@ struct mipi_dsi_host_ops {
  * @clk_lp2hs: Low Speed to High Speed Clock Transition Time
  */
 struct mipi_dsi_phy_timing {
-	u16 data_hs2lp;
-	u16 data_lp2hs;
+	u32 data_hs2lp;
+	u32 data_lp2hs;
 	u16 clk_hs2lp;
 	u16 clk_lp2hs;
 };
